[db] fix the Driver not found for jdbc:sqlite see #217
I don't know why but since the removal of the sqlite driver
when we launched the app with sqlite there is a driver not
found exception. It works with postgresql.
It works if we put the sqlite-jdbc jar *before* the datashare-db.jar.
Hypothesis : the driver self-registers itself at the class load time.
Sqlite driver is coming after the HikariPool so it is not seen when
trying to get the datasource pool.