feat: detect duplicate according to document id and path #1858
The report map is already in charge of avoiding a reindexing of the same document. For this reason, we consider a document to be a "Duplicate" of another document only if it has the same id (based on its hash) and a different path. This allows us to re-index an existing document while ensuring duplicates are detected.